[Akshita/Thoshiba] Alter.Logic This commit will alter testcase to find area of a square for decimal value

+++ src/main/java/com/thoughtworks/square/Square.java
@@ -2,12 +2,12 @@ package com.thoughtworks.square;
 
 public class Square {
 
-    private final int side;
-    public Square(int side) {
+    private final double side;
+    public Square(double side) {
         this.side = side;
     }
 
-    public int area() {
+    public double area() {
         if(side <= 0) return 0;
         return side * side;
     }
